Bulk Upload Feature is Implemented
Bulk upload feature is implemented and it is now available in the
component. The user first needs to upload an excel sheet with the list
of businesses along with their location coordinates. After that the user
can select one or more files by checking the corresponding checkbox.
After that by pressing the 'process file' button in the toolbar, those
selected file/files are processed and uploaded to the database.