Data Validation is added to the biz add form
At the moment the data validation for the biz add form is added. For biz
add bulk form the validation will be added in future as well.

The current implementation of validation check is using html5 standard
validation attributes and it is highlighting the fields for invalid data
when the form is submitted. In addition for required fields those fields
are also highlighted when the user leaves them without entering data.

In future will try to add message e.g. a balloon like tooltip to give
the appropriate message for the invalid data input.